ssl-utils: Build with -Wshorten-64-to-32
[...]/wireshark/epan/dissectors/packet-ssl-utils.c:3556:7: error: implicit
   conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'gint'
   (aka 'int') [-Werror,-Wshorten-64-to-32]
     i=rsa_decrypt_inplace(encrypted_pre_master->data_len,
      ~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Reviewed-on: https://code.wireshark.org/review/21976
Petri-Dish: Stig Bjørlykke <stig@bjorlykke.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>

+++ epan/dissectors/packet-ssl-utils.c
@@ -3530,7 +3530,7 @@ static gboolean
 ssl_decrypt_pre_master_secret(SslDecryptSession*ssl_session,
     StringInfo* encrypted_pre_master, gcry_sexp_t pk)
 {
-    gint i;
+    size_t i;
     char *err;
 
     if (!encrypted_pre_master)
@@ -3561,7 +3561,7 @@ ssl_decrypt_pre_master_secret(SslDecryptSession*ssl_session,
     }
 
     if (i!=48) {
-        ssl_debug_printf("%s wrong pre_master_secret length (%d, expected "
+        ssl_debug_printf("%s wrong pre_master_secret length (%zd, expected "
                          "%d)\n", G_STRFUNC, i, 48);
         return FALSE;
     }
